requirements.txt
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-SocketIO==2.4
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.11.10
aadict==0.2.2
argparse==1.2.1
asset==0.6.10
bcrypt==2.0.0
blinker==1.4
cffi==1.6.0
click==6.6
globre==0.1.4
itsdangerous==0.24
passwordmeter==0.1.8
pycparser==2.14
python-engineio==0.9.1
python-socketio==1.3
python-usernames==0.2.2
six==1.10.0
validate-email==1.3
wsgiref==0.1.2
eventlet==0.19.0
greenlet==0.4.9
